refactor all MockSender copies in the aggregator package
