+++ pkg/collector/corechecks/system/file_handles_test.go
@@ -47,8 +47,7 @@ func TestFhCheckLinux(t *testing.T) {
 	fileHandleCheck := new(fhCheck)
 	fileHandleCheck.Configure(nil, nil)
 
-	mock := new(MockSender)
-	aggregator.SetSender(mock, fileHandleCheck.ID())
+	mock := aggregator.NewMockSender(fileHandleCheck.ID())
 
 	mock.On("Gauge", "system.fs.file_handles.in_use", 0.008829499990145647, "", []string(nil)).Return().Times(1)
 	mock.On("Commit").Return().Times(1)
