+++ pkg/collector/corechecks/system/iostats_test.go
@@ -67,8 +67,7 @@ func TestIOCheck(t *testing.T) {
 	ioCheck := new(IOCheck)
 	ioCheck.Configure(nil, nil)
 
-	mock := new(MockSender)
-	aggregator.SetSender(mock, ioCheck.ID())
+	mock := aggregator.NewMockSender(ioCheck.ID())
 
 	expectedRates := 2
 	expectedGauges := 0
@@ -131,8 +130,7 @@ func TestIOCheckBlacklist(t *testing.T) {
 	ioCheck := new(IOCheck)
 	ioCheck.Configure(nil, nil)
 
-	mock := new(MockSender)
-	aggregator.SetSender(mock, ioCheck.ID())
+	mock := aggregator.NewMockSender(ioCheck.ID())
 
 	//set blacklist
 	bl, err := regexp.Compile("sd.*")
